+++ src/ossia/protocols/oscquery/oscquery_mirror_asio.cpp
@@ -10,7 +10,7 @@
 #include <ossia/network/osc/detail/osc_receive.hpp>
 #include <ossia/network/osc/detail/receiver.hpp>
 #include <ossia/network/osc/detail/sender.hpp>
-#include <ossia/network/websocket/client.hpp>
+#include <ossia/network/sockets/websocket_client.hpp>
 #include <ossia/network/http/http_client.hpp>
 #include <ossia/network/oscquery/detail/json_parser.hpp>
 #include <ossia/network/oscquery/detail/json_writer.hpp>
@@ -18,7 +18,7 @@
 #include <ossia/network/oscquery/detail/outbound_visitor.hpp>
 #include <ossia/network/oscquery/detail/value_to_json.hpp>
 #include <ossia/network/oscquery/detail/oscquery_protocol_common.hpp>
-#include <ossia/network/osc/detail/bidir.hpp>
+#include <ossia/network/sockets/udp_socket.hpp>
 #include <ossia/network/osc/detail/osc_1_1_extended_policy.hpp>
 #include <ossia/network/osc/detail/osc_protocol_common.hpp>
 #include <ossia/network/context.hpp>
